refactor: IssueTab for [ADR] Projects Tab Design & Implementation #6
